refactor: use SchemaEditorLite instead of SchemaEditorV1
